refactor: Request 메서드
