+++ Projects/Feature/Sources/Ranking/ViewModel/RankingDrinkViewModel.swift
@@ -10,68 +10,27 @@ import Foundation
 import Service
 
 final class RankingDrinkViewModel {
-    
     private let jsonDecoder = JSONDecoder()
     
-    var cancelBag = Set<AnyCancellable>()
-    var drinkDatasource: [Ranking] = []
-    var combinationDatasource: [Ranking] = []
-    var rankingBasePublisher: AnyPublisher<[Ranking], Never> {
-        return rankingBaseSubject.eraseToAnyPublisher()
-    }
-    var rankingDrinkPublisher: AnyPublisher<[Ranking], Never> {
-        return rankingDrinkSubject.eraseToAnyPublisher()
-    }
-    var rankingCombinationPublisher: AnyPublisher<[Ranking], Never> {
-        return rankingCombinationSubject.eraseToAnyPublisher()
-    }
-    
-    private var rankingBaseSubject = PassthroughSubject<[Ranking], Never>()
+    private var cancelBag = Set<AnyCancellable>()
     private var rankingDrinkSubject = PassthroughSubject<[Ranking], Never>()
-    private var rankingCombinationSubject = PassthroughSubject<[Ranking], Never>()
+    private var drinkDatasource: [Ranking] = []
     
-    init() {
-        requestRankingAlcohol()
-        requestRankingCombination()
-    }
+    init() {}
     
     public func requestRankingAlcohol() {
         NetworkWrapper.shared.getBasicTask(stringURL: "/ranks/alcohol") { [weak self] result in
             guard let self = self else { return }
             switch result {
-            case .success(let responseData):
-                do {
-                    let data = try self.jsonDecoder.decode(Ranking.self, from: responseData)
+            case .success(let response):
+                if let data = try? self.jsonDecoder.decode(Ranking.self, from: response) {
+                    print(data)
                     self.drinkDatasource = [data]
                     self.rankingDrinkSubject.send(self.drinkDatasource)
-                    self.rankingBaseSubject.send(self.drinkDatasource)
-                    print("디코딩 성공: \(self.drinkDatasource)")
-                } catch {
-                    print("디코딩 에러: \(error)")
                 }
             case .failure(let error):
-                print(error)
+                print("RankingDrinkViewModel.requestRankingAlcohol(): \(error)")
             }
-        }
-    }
-    
-    public func requestRankingCombination() {
-        NetworkWrapper.shared.getBasicTask(stringURL: "/ranks/combinations") { [weak self] result in
-            guard let self = self else { return }
-            switch result {
-            case .success(let responseData):
-                do {
-                    let data = try self.jsonDecoder.decode(Ranking.self, from: responseData)
-                    self.combinationDatasource = [data]
-                    self.rankingCombinationSubject.send(self.combinationDatasource)
-                    print("디코딩 성공: \(self.combinationDatasource)")
-                } catch {
-                    print("디코딩 에러: \(error)")
-                }
-            case .failure(let error):
-                print(error)
-            }
-            
         }
     }
     
@@ -79,7 +38,11 @@ final class RankingDrinkViewModel {
         return drinkDatasource.first?.ranking?.count ?? 0
     }
     
-    public func combinationDatasourceCount() -> Int {
-        return combinationDatasource.first?.ranking?.count ?? 0
+    public func getDrinkDatasource(to index: IndexPath) -> RankingItem? {
+        return drinkDatasource.first?.ranking?[index.item]
+    }
+    
+    public var rankingDrinkPublisher: AnyPublisher<[Ranking], Never> {
+        return rankingDrinkSubject.eraseToAnyPublisher()
     }
 }
